implement new build image that has clang-19 and cmake-31 (latest)

+++ scripts/ros/ros-2-prep.sh
@@ -1,10 +1,12 @@
 #!/usr/bin/env bash
+set -e
+set -x
 
-script_dir=$(dirname $0)
+script_dir="$(dirname $0)"
 
 locale  # check for UTF-8
 
-sudo apt update && sudo apt install locales
+sudo apt update && sudo apt install -y locales
 sudo locale-gen en_US en_US.UTF-8
 sudo update-locale LC_ALL=en_US.UTF-8 LANG=en_US.UTF-8
 export LANG=en_US.UTF-8
@@ -14,20 +16,31 @@ locale  # verify settings
 sudo apt install -y software-properties-common
 sudo add-apt-repository universe
 
-sudo apt update && sudo apt install curl gnupg2 lsb-release
-sudo curl -sSL https://raw.githubusercontent.com/ros/rosdistro/master/ros.key  -o /usr/share/keyrings/ros-archive-keyring.gpg
+sudo apt update && sudo apt install -y wget curl gnupg2 lsb-release ca-certificates
 
-echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/ros-archive-keyring.gpg] http://packages.ros.org/ros2/ubuntu $(source /etc/os-release && echo $UBUNTU_CODENAME) main" | sudo tee /etc/apt/sources.list.d/ros2.list > /dev/null
+# Clang 19
+CLANG_VERSION=19
+curl -sSL https://apt.llvm.org/llvm.sh -o "$script_dir/llvm.sh"
+chmod +x "$script_dir/llvm.sh"
+sudo "$script_dir/llvm.sh" $CLANG_VERSION all
+sudo "$script_dir/update-alternatives-clang.sh" $CLANG_VERSION 1
+
+
+if [[ ! (-f /etc/apt/sources.list.d/ros2-latest.list || -f /etc/apt/sources.list.d/ros2.list) ]]; then
+  # sudo rm /etc/apt/sources.list.d/ros2-latest.list # remove old ros2-latest.list from the docker image
+  sudo curl -sSL https://raw.githubusercontent.com/ros/rosdistro/master/ros.key  -o /usr/share/keyrings/ros-archive-keyring.gpg
+  echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/ros-archive-keyring.gpg] http://packages.ros.org/ros2/ubuntu $(source /etc/os-release && echo $UBUNTU_CODENAME) main" | sudo tee /etc/apt/sources.list.d/ros2.list > /dev/null
+fi
 
 # CMake
-wget -O - https://apt.kitware.com/keys/kitware-archive-latest.asc 2>/dev/null | gpg --dearmor - | sudo tee /etc/apt/trusted.gpg.d/kitware.gpg >/dev/null
-sudo apt-add-repository 'deb https://apt.kitware.com/ubuntu/ $(lsb_release -cs) main'
+test -f /usr/share/doc/kitware-archive-keyring/copyright || wget -O - https://apt.kitware.com/keys/kitware-archive-latest.asc 2>/dev/null | gpg --dearmor - | sudo tee /usr/share/keyrings/kitware-archive-keyring.gpg >/dev/null
+echo 'deb [signed-by=/usr/share/keyrings/kitware-archive-keyring.gpg] https://apt.kitware.com/ubuntu/ jammy main' | sudo tee /etc/apt/sources.list.d/kitware.list >/dev/null
 
 sudo apt update
+test -f /usr/share/doc/kitware-archive-keyring/copyright || sudo rm /usr/share/keyrings/kitware-archive-keyring.gpg
 sudo apt install kitware-archive-keyring
-sudo rm /etc/apt/trusted.gpg.d/kitware.gpg
 
-sudo apt upgrade
+sudo apt upgrade -y
 
 sudo apt update && sudo apt install -y \
   build-essential \
@@ -41,8 +54,8 @@ sudo apt update && sudo apt install -y \
   python3-pytest-cov \
   python3-rosdep \
   python3-setuptools \
-  python3-vcstool \
-  wget
+  python3-vcstool
+
 # install some pip packages needed for testing
 python3 -m pip install -U \
   argcomplete \
@@ -71,13 +84,6 @@ sudo rm -f /etc/ros/rosdep/sources.list.d/*
 sudo rosdep init
 rosdep update
 
-sudo apt install ros-humble-desktop
+sudo apt install -y ros-humble-desktop
 # sudo apt install ros-humble-ros-base
-sudo apt install ros-dev-tools
-
-# Clang 19
-CLANG_VERSION=19
-wget https://apt.llvm.org/llvm.sh
-chmod +x llvm.sh
-sudo ./llvm.sh $CLANG_VERSION all
-sudo "$script_dir/update-alternatives-clang.sh" $CLANG_VERSION 1
+sudo apt install -y ros-dev-tools
